Moved CheckBoxAccessibleObject to own file
