+++ src/System.Windows.Forms/src/System/Windows/Forms/CheckBox.cs
@@ -23,7 +23,7 @@ namespace System.Windows.Forms
     [DefaultBindingProperty(nameof(CheckState))]
     [ToolboxItem("System.Windows.Forms.Design.AutoSizeToolboxItem," + AssemblyRef.SystemDesign)]
     [SRDescription(nameof(SR.DescriptionCheckBox))]
-    public class CheckBox : ButtonBase
+    public partial class CheckBox : ButtonBase
     {
         private static readonly object EVENT_CHECKEDCHANGED = new object();
         private static readonly object EVENT_CHECKSTATECHANGED = new object();
@@ -644,85 +644,6 @@ namespace System.Windows.Forms
             // We shouldn't need to convert the enum to int
             int checkState = (int)CheckState;
             return s + ", CheckState: " + checkState.ToString(CultureInfo.InvariantCulture);
-        }
-
-        public class CheckBoxAccessibleObject : ButtonBaseAccessibleObject
-        {
-            public CheckBoxAccessibleObject(Control owner) : base(owner)
-            {
-            }
-
-            public override string DefaultAction
-            {
-                get
-                {
-                    string defaultAction = Owner.AccessibleDefaultActionDescription;
-                    if (defaultAction != null)
-                    {
-                        return defaultAction;
-                    }
-
-                    if (((CheckBox)Owner).Checked)
-                    {
-                        return SR.AccessibleActionUncheck;
-                    }
-                    else
-                    {
-                        return SR.AccessibleActionCheck;
-                    }
-                }
-            }
-
-            public override AccessibleRole Role
-            {
-                get
-                {
-                    AccessibleRole role = Owner.AccessibleRole;
-                    if (role != AccessibleRole.Default)
-                    {
-                        return role;
-                    }
-                    return AccessibleRole.CheckButton;
-                }
-            }
-
-            public override AccessibleStates State
-            {
-                get
-                {
-                    switch (((CheckBox)Owner).CheckState)
-                    {
-                        case CheckState.Checked:
-                            return AccessibleStates.Checked | base.State;
-                        case CheckState.Indeterminate:
-                            return AccessibleStates.Indeterminate | base.State;
-                    }
-
-                    return base.State;
-                }
-            }
-
-            public override void DoDefaultAction()
-            {
-                CheckBox cb = Owner as CheckBox;
-
-                if (cb != null)
-                {
-                    cb.AccObjDoDefaultAction = true;
-                }
-
-                try
-                {
-                    base.DoDefaultAction();
-                }
-                finally
-                {
-                    if (cb != null)
-                    {
-                        cb.AccObjDoDefaultAction = false;
-                    }
-                }
-            }
         }
     }
 }
